refactor: 패키지 구조 변경
* driver 패키지 추가 및 관련 파일 옮김
