+++ BE/src/main/java/ifive/idrop/controller/ParentController.java
@@ -4,10 +4,10 @@ import ifive.idrop.dto.request.DriverListRequest;
 import ifive.idrop.dto.response.CurrentPickUpResponse;
 import ifive.idrop.dto.response.DriverListResponse;
 import ifive.idrop.dto.response.ParentSubscribeInfoResponse;
-import ifive.idrop.entity.Driver;
+import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.exception.CommonException;
 import ifive.idrop.exception.ErrorCode;
-import ifive.idrop.service.DriverService;
+import ifive.idrop.driver.service.DriverService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
