+++ BE/src/main/java/ifive/idrop/dto/request/SignUpRequest.java
@@ -1,6 +1,6 @@
 package ifive.idrop.dto.request;
 
-import ifive.idrop.entity.Driver;
+import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.entity.Parent;
 import ifive.idrop.entity.Users;
 import ifive.idrop.entity.enums.Role;
