+++ BE/src/main/java/ifive/idrop/entity/PickUp.java
@@ -1,5 +1,6 @@
 package ifive.idrop.entity;
 
+import ifive.idrop.driver.domain.Driver;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
